edit much

+++ app/Http/Livewire/Sertifikasi/Modal/EditDokumen.php
@@ -12,7 +12,7 @@ use Illuminate\Support\Str;
 class EditDokumen extends ModalComponent
 {
     use WithFileUploads;
-    public $doc, $doc_id, $ruas, $document, $data, $nama_dokumen;
+    public $doc, $doc_id, $ruas, $document, $data, $nama_dokumen, $nama_dokumen_edited = [];
 
     public function mount($id, $data)
     {
@@ -45,34 +45,54 @@ class EditDokumen extends ModalComponent
             $q->where('registrations.id', $this->ruas);
         }])->findOrFail($id);
         $bujt = $doc->registration[0];
+        dd($bujt);
 
         if ($doc->type == 'file') {
-            $this->validate([
-                'nama_dokumen' => 'mimes:pdf|max:102400',
-            ], [
-                'nama_dokumen.max' => 'Dokumen harus berukuran maksimal 15MB',
-                'nama_dokumen.mimes' => 'Dokumen harus berbentuk PDF',
-            ]);
 
-            $file = $this->nama_dokumen;
-            preg_match("/(?:\w+(?:\W+|$)){0,10}/", $doc->nama_dokumen, $matches);
+            if ($bujt->pivot->status == 2) {
+                // $this->validate([
+                //     'nama_dokumen' => 'required',
+                //     'nama_dokumen.*' => 'mimes:pdf|max:102400',
+                // ], [
+                //     'nama_dokumen.*.max' => 'Dokumen harus berukuran maksimal 15MB',
+                //     'nama_dokumen.*.mimes' => 'Dokumen harus berbentuk PDF',
+                // ]);
+                // # code...
+                // $i = 1;
+                // foreach ($this->nama_dokumen_edited as $data) {
+                //     preg_match("/(?:\w+(?:\W+|$)){0,10}/", $doc->nama_dokumen, $matches);
 
-            $nama_file = Str::slug($bujt->nama_ruas) . '-' . Str::slug($doc->kode) . '-' . Str::slug($matches[0]);
-            $data = $nama_file . '.' . $file->extension();
+                //     $nama_file = Str::slug($bujt->nama_ruas) . '-' . Str::slug($doc->kode) . '-' . Str::slug($matches[0]);
 
-            $path = 'storage/checklist-dokumen/' . $bujt->nama_bujt . '/' . $bujt->nama_ruas;
-            $filename = $path  . '/' . $data;
+                //     # code...
+                // }
+            } else {
+                $this->validate([
+                    'nama_dokumen' => 'mimes:pdf|max:102400',
+                ], [
+                    'nama_dokumen.max' => 'Dokumen harus berukuran maksimal 15MB',
+                    'nama_dokumen.mimes' => 'Dokumen harus berbentuk PDF',
+                ]);
+                $file = $this->nama_dokumen;
+                preg_match("/(?:\w+(?:\W+|$)){0,10}/", $doc->nama_dokumen, $matches);
 
-            if (file_exists($filename)) {
-                unlink($filename);
+                $nama_file = Str::slug($bujt->nama_ruas) . '-' . Str::slug($doc->kode) . '-' . Str::slug($matches[0]);
+                $data = $nama_file . '.' . $file->extension();
+
+                $path = 'storage/checklist-dokumen/' . $bujt->nama_bujt . '/' . $bujt->nama_ruas;
+                $filename = $path  . '/' . $data;
+
+                if (file_exists($filename)) {
+                    unlink($filename);
+                }
+
+                $bujt->pivot->status = 1;
+                $bujt->pivot->nama_dokumen = $data;
+                $bujt->pivot->save();
+
+                $file->storeAs('checklist-dokumen/' . $bujt->nama_bujt . '/' . $bujt->nama_ruas, $data);
+                $this->nama_dokumen = null;
             }
-
-            $bujt->pivot->status = 1;
-            $bujt->pivot->nama_dokumen = $data;
-            $bujt->pivot->save();
-
-            $file->storeAs('checklist-dokumen/' . $bujt->nama_bujt . '/' . $bujt->nama_ruas, $data);
-            $this->nama_dokumen = null;
         } else {
             $this->validate([
                 'nama_dokumen' => 'required',
@@ -83,6 +103,8 @@ class EditDokumen extends ModalComponent
             $bujt->pivot->status = 1;
             $bujt->pivot->nama_dokumen = $this->nama_dokumen;
             $bujt->pivot->save();
+
+            $this->nama_dokumen = null;
         }
 
         $this->dispatchBrowserEvent(
